+++ src/graphiql/GraphiQL.ts
@@ -1,0 +1,5 @@
+export class GraphiQL {
+  static path() {
+    return "node_modules/graphqlade/dist/graphiql/public";
+  }
+}