+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from "./client";
 export * from "./codegen";
+export * from "./graphiql";
 export * from "./http";
 export * from "./introspect";
 export * from "./read";